+++ zia/resource_zia_dlp_dictionaries_test.go
@@ -45,6 +45,12 @@ func TestAccResourceDLPDictionariesBasic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceTypeAndName, "patterns.#", "2"),
 				),
 			},
+			// Import test
+			{
+				ResourceName:      resourceTypeAndName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
 		},
 	})
 }