+++ zia/resource_zia_dlp_engines_test.go
@@ -43,6 +43,12 @@ func TestAccResourceDLPEnginesBasic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceTypeAndName, "custom_dlp_engine", strconv.FormatBool(variable.DLPCustomEngine)),
 				),
 			},
+			// Import test
+			{
+				ResourceName:      resourceTypeAndName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
 		},
 	})
 }