+++ zia/resource_zia_fw_filtering_network_services_groups_test.go
@@ -43,6 +43,12 @@ func TestAccResourceFWNetworkServiceGroupsBasic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceTypeAndName, "services.#", "1"),
 				),
 			},
+			// Import test
+			{
+				ResourceName:      resourceTypeAndName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
 		},
 	})
 }