+++ zia/resource_zia_traffic_forwarding_static_ips_test.go
@@ -47,6 +47,12 @@ func TestAccResourceTrafficForwardingStaticIPBasic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceTypeAndName, "geo_override", strconv.FormatBool(variable.StaticGeoOverride)),
 				),
 			},
+			// Import test
+			{
+				ResourceName:      resourceTypeAndName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
 		},
 	})
 }
@@ -109,7 +115,7 @@ func testAccCheckTrafficForwardingStaticIPExists(resource string, rule *staticip
 func testAccCheckTrafficForwardingStaticIPConfigure(resourceTypeAndName, generatedName, ipAddress string, routableIP, geoOverride bool) string {
 	return fmt.Sprintf(`
 
-// location management resource
+// static ip resource
 %s
 
 data "%s" "%s" {