+++ docs/guides/release-notes.md
@@ -12,9 +12,22 @@ description: |-
 Track all ZIA Terraform provider's releases. New resources, features, and bug fixes will be tracked here.
 
 ---
-``Last updated: v2.6.3``
+``Last updated: v2.7.0``
 
 ---
+
+## 2.7.0 (November, xx 2023)
+
+### Notes
+
+- Release date: **(November, xx 2023)**
+- Supported Terraform version: **v1.x**
+
+### Enhancements
+
+NEW - RESOURCES, DATA SOURCES
+
+- [PR #283](https://github.com/zscaler/terraform-provider-zia/pull/283) - Added support to ZPA ZPA Application Segment Traffic (ZPA Gateway)feature
 
 ## 2.6.3 (October, 18 2023)